refactor(Models): split models file for sanity of the man
